requirements.txt
Flask==3.0.3
gunicorn==21.2.0
mysql-connector-python==8.3.0
numpy==1.26.4
opencv-contrib-python-headless==4.9.0.80
pandas==2.2.2
reportlab==4.2.0
openpyxl==3.1.2
Werkzeug==3.0.3